add[FE]: Add API to following page DM button(#378)
