+++ frontend/src/components/organisms/Following.tsx
@@ -1,3 +1,6 @@
+import { useNavigate } from 'react-router-dom';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { getDmChannel, unfollowUser } from 'api/api.v1';
 import UserList from 'components/molecule/UserList';
 import Button from 'components/atoms/Button';
 import GameInviteButton from 'components/molecule/GameInviteButton';
@@ -5,10 +8,36 @@ import { UserType } from 'types';
 
 interface Props {
   users: UserType[];
-  onClickUnfollow: (e: React.MouseEvent<HTMLElement>) => void;
 }
 
-export default function Following({ users, onClickUnfollow }: Props) {
+export default function Following({ users }: Props) {
+  const navigate = useNavigate();
+  const queryClient = useQueryClient();
+
+  const getDmChannelMutation = useMutation({
+    mutationFn: getDmChannel,
+    onSuccess: (data) => navigate(`/channel/${data.id}`),
+  });
+
+  const unfollowUserMutation = useMutation({
+    mutationFn: unfollowUser,
+    onSuccess: () => queryClient.invalidateQueries(['getMyFollowing']),
+  });
+
+  const handleClickDm = (e: React.MouseEvent<HTMLElement>) => {
+    const ancestorElement = e.currentTarget.closest('[data-user-id]');
+    if (!(ancestorElement instanceof HTMLElement)) return;
+    const userId = ancestorElement.dataset.userId;
+    getDmChannelMutation.mutate(Number(userId));
+  };
+
+  const handleClickUnfollow = (e: React.MouseEvent<HTMLElement>) => {
+    const ancestorElement = e.currentTarget.closest('[data-user-id]');
+    if (!(ancestorElement instanceof HTMLElement)) return;
+    const userId = ancestorElement.dataset.userId;
+    unfollowUserMutation.mutate(Number(userId));
+  };
+
   return (
     <section className="section">
       <h2 className="section-title">팔로잉</h2>
@@ -18,12 +47,12 @@ export default function Following({ users, onClickUnfollow }: Props) {
         hasStatus={true}
         buttons={[
           <GameInviteButton key="button0" />,
-          <Button key="button1" primary size="small">
+          <Button onClick={handleClickDm} key="button1" primary size="small">
             메시지 보내기
           </Button>,
           <Button
             key="button2"
-            onClick={onClickUnfollow}
+            onClick={handleClickUnfollow}
             linkStyle
             className="text-red"
             size="small"
